src/functions/requirements.txt
google-cloud-dataproc
google-cloud-bigquery